requirements.txt
Django==4.2.1
djangorestframework==3.14.0
djoser==2.2.0
djangorestframework_simplejwt==5.2.2